Icons: Add more key words #1238
