+++ package.json
@@ -49,6 +49,7 @@
   "sass": "^1.38.1"
  },
  "dependencies": {
+  "csv-parse": "^5.3.6",
   "markdown-it": "^12.3.2"
  }
 }
